data-engineering/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sho-da-igram"
version = "0.1.0"
description = "Data Engineering Pipeline for Game Recommendations"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Teodor Angeleski", email = "teodorangeleski@gmail.com"}
]

dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pre-commit>=3.4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sho_da_igram"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
